Remove quick Rotator
